+++ lib/molecules/search_bar/components/zwap_seach_bar_filter_widget.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ZwapSearchBarFilterWidget extends StatelessWidget {
+  
+
+  const ZwapSearchBarFilterWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}